+++ src/pcb_spec/calc/__init__.py
@@ -1,0 +1,1 @@
+"""PCB calculation tools: impedance, current capacity."""