+++ src/pcb_spec/emit/__init__.py
@@ -1,0 +1,1 @@
+"""EDA rule translators: KiCad .kicad_dru emitter."""